fix from pr: core: renamed and privatize indexed port collection
